Support documentation language families
Implement latest concepts introduced in Racket 9.1.

As discussed in <https://github.com/racket/racket/pull/5409>.

- racket-hash-lang: Expose get-info documentation-language-family.

- Expose get-main-language and get-language-families in new
doc-families back end command. Return a cons of the main family name,
and, a list of association lists with details for each family. The
back end normalizes some variations in data, as well as changing
abstract module paths into concrete absolute real paths.

- Extract language family from documentation page. Display in header
line. Supply racket-describe-choose-language-family to pick a family
for "navigating as", akin to web browser, which ca affect "top" and
"up" navigation, as well as sorting search candidates. The command is
bound to a key, as well as button-ized family text in the header line.

- Implement racket-describe-about-language-family, to view the doc
page (if any) describing the family (often a "Guide" top page).

- describe-search: Sort by the one "preferred" family first, and then
sort by other families that an index item may supply, and using the
'order key from get-language-families. This means for example that
search candidates for HtDP will appear /after/ those for Rhombus, when
the preferred family is neither.

- Fix bug extracting prev/up/next links. (Before, when no prev link,
we'd mismatch up and next.) Strictly speaking this is independent of
doc lang families, but I didn't preserve it as a distinct commit.